clean up `emit_drop_facts`
- remove dependency on `TypeChecker`
- move to legacy fact generation module
- remove polonius module from liveness
